renames actions/index to actions/docs
